Moved MagicMock and the install_reactor setup_class code to testutils
