+++ tests/golem/envs/test_localhost_env.py
@@ -9,12 +9,10 @@ from golem_task_api import (
 from golem_task_api.structs import Subtask
 from grpclib.exceptions import StreamTerminatedError
 from twisted.internet.defer import inlineCallbacks
-from twisted.trial.unittest import TestCase as TwistedTestCase
 
-from golem.core.common import install_reactor
 from golem.core.deferred import deferred_from_future
 from golem.task.task_api import EnvironmentTaskApiService
-from golem.tools.testwithreactor import uninstall_reactor
+from golem.tools.testwithasyncio import TwistedAsyncioTestCase
 from tests.golem.envs.localhost import (
     LocalhostEnvironment,
     LocalhostConfig,
@@ -23,19 +21,7 @@ from tests.golem.envs.localhost import (
 )
 
 
-class TestLocalhostEnv(TwistedTestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        try:
-            uninstall_reactor()  # Because other tests don't clean up
-        except AttributeError:
-            pass
-        install_reactor()
-
-    @classmethod
-    def tearDownClass(cls):
-        uninstall_reactor()
+class TestLocalhostEnv(TwistedAsyncioTestCase):
 
     @staticmethod
     def _get_service(prereq: LocalhostPrerequisites) -> TaskApiService:
